refactor(ngcc): expose logging level on the logger (#35861)
PR Close #35861
