+++ packages/compiler-cli/ngcc/index.ts
@@ -8,8 +8,8 @@
 import {CachedFileSystem, NodeJSFileSystem, setFileSystem} from '../src/ngtsc/file_system';
 
 import {AsyncNgccOptions, NgccOptions, SyncNgccOptions, mainNgcc} from './src/main';
-export {ConsoleLogger, LogLevel} from './src/logging/console_logger';
-export {Logger} from './src/logging/logger';
+export {ConsoleLogger} from './src/logging/console_logger';
+export {LogLevel, Logger} from './src/logging/logger';
 export {AsyncNgccOptions, NgccOptions, SyncNgccOptions} from './src/main';
 export {PathMappings} from './src/utils';
