+++ packages/compiler-cli/ngcc/src/main.ts
@@ -31,8 +31,8 @@ import {LockFileAsync, LockFileSync} from './execution/lock_file';
 import {SingleProcessExecutorAsync, SingleProcessExecutorSync} from './execution/single_process_executor';
 import {ParallelTaskQueue} from './execution/task_selection/parallel_task_queue';
 import {SerialTaskQueue} from './execution/task_selection/serial_task_queue';
-import {ConsoleLogger, LogLevel} from './logging/console_logger';
-import {Logger} from './logging/logger';
+import {ConsoleLogger} from './logging/console_logger';
+import {LogLevel, Logger} from './logging/logger';
 import {hasBeenProcessed} from './packages/build_marker';
 import {NgccConfiguration} from './packages/configuration';
 import {EntryPoint, EntryPointJsonProperty, EntryPointPackageJson, SUPPORTED_FORMAT_PROPERTIES, getEntryPointFormat} from './packages/entry_point';
