+++ packages/compiler-cli/ngcc/test/helpers/mock_logger.ts
@@ -6,10 +6,12 @@
  * found in the LICENSE file at https://angular.io/license
  */
 
-import {Logger} from '../../src/logging/logger';
+import {LogLevel, Logger} from '../../src/logging/logger';
 
 export class MockLogger implements Logger {
-  logs: {[P in keyof Logger]: string[][]} = {
+  constructor(public level = LogLevel.info) {}
+
+  logs: {[P in Exclude<keyof Logger, 'level'>]: string[][]} = {
     debug: [],
     info: [],
     warn: [],
