+++ packages/compiler-cli/ngcc/test/logging/console_logger_spec.ts
@@ -5,7 +5,8 @@
  * Use of this source code is governed by an MIT-style license that can be
  * found in the LICENSE file at https://angular.io/license
  */
-import {ConsoleLogger, DEBUG, ERROR, LogLevel, WARN} from '../../src/logging/console_logger';
+import {ConsoleLogger, DEBUG, ERROR, WARN} from '../../src/logging/console_logger';
+import {LogLevel} from '../../src/logging/logger';
 
 describe('ConsoleLogger', () => {
   it('should pass through calls to Console', () => {
@@ -46,5 +47,12 @@ describe('ConsoleLogger', () => {
 
     logger.error('error', 'test');
     expect(console.error).toHaveBeenCalledWith(ERROR, 'error', 'test');
+  });
+
+  it('should expose the logging level', () => {
+    expect(new ConsoleLogger(LogLevel.debug).level).toEqual(LogLevel.debug);
+    expect(new ConsoleLogger(LogLevel.info).level).toEqual(LogLevel.info);
+    expect(new ConsoleLogger(LogLevel.warn).level).toEqual(LogLevel.warn);
+    expect(new ConsoleLogger(LogLevel.error).level).toEqual(LogLevel.error);
   });
 });
